Otbasy Bank_RAG with LLM/requirements.txt
Flask
langchain
langchain-openai
langchain-community
python-dotenv
